% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytorf.R
\name{cytorf}
\alias{cytorf}
\title{Perform unsupervised clusering using random forest}
\usage{
cytorf(X, num.trees = 50, scale = 5, seed = 1234)
}
\description{
Perform unsupervised clusering using random forest
}
