% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytorf.ui.R
\name{cytorf.ui}
\alias{cytorf.ui}
\title{Shiny interface to CytoRF clustering algorithm}
\usage{
cytorf.ui(port = 1234)
}
\arguments{
\item{port}{Numeric port number for shiny server. Defaults to 1234}
}
\description{
Shiny interface to CytoRF clustering algorithm
}
